refactor(user_service): move DTOs to web/dto
