+++ src/main/java/com/renewsim/backend/user_service/application/mapper/UserServiceMapper.java
@@ -1,8 +1,8 @@
 package com.renewsim.backend.user_service.application.mapper;
 
 import com.renewsim.backend.user_service.domain.model.User;
-import com.renewsim.backend.user_service.dto.UserCreateRequest;
-import com.renewsim.backend.user_service.dto.UserResponse;
+import com.renewsim.backend.user_service.web.dto.UserCreateRequest;
+import com.renewsim.backend.user_service.web.dto.UserResponse;
 import com.renewsim.backend.user_service.infrastructure.persistence.entity.UserEntity;
 import com.renewsim.backend.shared.domain.vo.RoleName;
 import org.mapstruct.Mapper;
