+++ src/main/java/com/renewsim/backend/user_service/application/service/CreateUserService.java
@@ -11,8 +11,8 @@ import com.renewsim.backend.user_service.application.port.in.CreateUserUseCase;
 import com.renewsim.backend.user_service.application.port.out.UserRepositoryPort;
 import com.renewsim.backend.user_service.domain.model.User;
 import com.renewsim.backend.user_service.domain.service.UserPolicy;
-import com.renewsim.backend.user_service.dto.UserCreateRequest;
-import com.renewsim.backend.user_service.dto.UserResponse;
+import com.renewsim.backend.user_service.web.dto.UserCreateRequest;
+import com.renewsim.backend.user_service.web.dto.UserResponse;
 import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 import com.renewsim.backend.shared.domain.vo.RoleName;
