+++ src/main/java/com/renewsim/backend/user_service/application/service/GetUserService.java
@@ -10,7 +10,7 @@ import com.renewsim.backend.shared.exception.UserNotFoundException;
 import com.renewsim.backend.user_service.application.port.in.GetUserUseCase;
 import com.renewsim.backend.user_service.application.port.out.UserRepositoryPort;
 import com.renewsim.backend.user_service.domain.model.User;
-import com.renewsim.backend.user_service.dto.UserResponse;
+import com.renewsim.backend.user_service.web.dto.UserResponse;
 import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 
 @Slf4j
