+++ src/main/java/com/renewsim/backend/user_service/web/dto/UserResponse.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.user_service.dto;
+package com.renewsim.backend.user_service.web.dto;
 
 import java.time.Instant;
 import java.util.Set;
