+++ src/test/java/com/renewsim/backend/user_service/application/service/CreateUserServiceTest.java
@@ -4,8 +4,8 @@ import com.renewsim.backend.shared.domain.vo.RoleName;
 import com.renewsim.backend.shared.exception.UserAlreadyExistsException;
 import com.renewsim.backend.user_service.application.port.out.UserRepositoryPort;
 import com.renewsim.backend.user_service.domain.model.User;
-import com.renewsim.backend.user_service.dto.UserCreateRequest;
-import com.renewsim.backend.user_service.dto.UserResponse;
+import com.renewsim.backend.user_service.web.dto.UserCreateRequest;
+import com.renewsim.backend.user_service.web.dto.UserResponse;
 import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
