+++ src/test/java/com/renewsim/backend/user_service/infrastructure/mapper/UserMapperTest.java
@@ -4,8 +4,8 @@ import com.renewsim.backend.shared.domain.vo.RoleName;
 import com.renewsim.backend.role_service.infrastructure.persistence.entity.RoleEntity;
 import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 import com.renewsim.backend.user_service.domain.model.User;
-import com.renewsim.backend.user_service.dto.UserCreateRequest;
-import com.renewsim.backend.user_service.dto.UserResponse;
+import com.renewsim.backend.user_service.web.dto.UserCreateRequest;
+import com.renewsim.backend.user_service.web.dto.UserResponse;
 import com.renewsim.backend.user_service.infrastructure.persistence.entity.UserEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
