+++ src/test/java/com/renewsim/backend/user_service/it/UserControllerIT.java
@@ -2,7 +2,7 @@ package com.renewsim.backend.user_service.it;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.renewsim.backend.shared.exception.GlobalExceptionHandler;
-import com.renewsim.backend.user_service.dto.UserCreateRequest;
+import com.renewsim.backend.user_service.web.dto.UserCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
